+++ source/texk/web2c/luatexdir/lua/texluac.c
@@ -156,9 +156,9 @@ static const Proto *combine(lua_State * L, int n)
         f->source = luaS_newliteral(L, "=(" PROGNAME ")");
         f->maxstacksize = 1;
         pc = 2 * n + 1;
-        f->code = luaM_newvector(L, pc, Instruction);
+        f->code = luaM_newvector(L, (unsigned long)pc, Instruction);
         f->sizecode = pc;
-        f->p = luaM_newvector(L, n, Proto *);
+        f->p = luaM_newvector(L, (unsigned long)n, Proto *);
         f->sizep = n;
         pc = 0;
         for (i = 0; i < n; i++) {
